build: refactor changed handler and added the file directory options for commands and events
