+++ src/events/interactionCreate.ts
@@ -1,7 +1,6 @@
 import { Client, Events, Interaction } from 'discord.js';
-import { loadSlashCommands } from '../modules/bot/slashCommandsLoader';
 
-const slashCommands = loadSlashCommands();
+import { loader } from '../index';
 
 export default {
   name: Events.InteractionCreate,
@@ -9,7 +8,9 @@ export default {
   async execute(interaction: Interaction, client: Client) {
     if (!interaction.isChatInputCommand()) return;
     const { commandName } = interaction;
-    const slashCommand = (await slashCommands).get(commandName);
+
+    const slashCommand = (await loader)!.commands.get(commandName);
+    console.log(slashCommand);
     slashCommand?.execute(interaction, client);
   },
 };
