+++ src/events/threadCreate.ts
@@ -1,5 +1,5 @@
 import { Events, ThreadChannel } from 'discord.js';
-import { cohereAIHandler } from './handlers/cohereAIHandler';
+import { cohereAIHandler } from '../modules/utils/cohereAIHandler';
 
 export default {
   name: Events.ThreadCreate,
